+++ backend/canisters/local_user_index/impl/src/model/mod.rs
@@ -1,5 +1,7 @@
 pub mod bots_map;
 pub mod global_user_map;
+pub mod local_community_map;
+pub mod local_group_map;
 pub mod local_user_map;
 pub mod referral_codes;
 pub mod user_event_batch;